backend_django/lolBTI/requirements.txt
asgiref==3.2.10
branca==0.3.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==3.1
djangorestframework==3.11.1
folium==0.10.1
idna==2.8
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.17.3
pandas==0.25.3
pip==18.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scipy==1.3.1
seaborn==0.9.0
setuptools==40.6.2
six==1.13.0
sqlparse==0.3.1
tqdm==4.48.2
urllib3==1.25.7